requirements.txt
appnope==0.1.3
asttokens==2.0.5
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
clickhouse-driver==0.2.4
cryptography==36.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
duckdb==0.5.1
entrypoints==0.4
envyaml==1.10.211231
executing==0.8.3
fastjsonschema==2.16.2
greenlet==1.1.2
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-client==7.2.2
jupyter-core==4.11.2
jupyterlab-pygments==0.2.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==2.0.4
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==2.5.2
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml-env==1.1.5
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
simple-salesforce==1.11.6
six==1.16.0
slack-sdk==3.4.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
stack-data==0.2.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.4.0
typer==0.6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
yarg==0.1.9
zeep==4.1.0
